Food ordering Java working setup
