+++ java/food-ordering/app/restate-app/src/main/java/dev/restate/sdk/examples/types/RestaurantRequest.java
