+++ java/food-ordering/app/restate-app/src/main/java/dev/restate/sdk/examples/utils/TypeUtils.java
@@ -1,0 +1,14 @@
+package dev.restate.sdk.examples.utils;
+
+
+import dev.restate.sdk.examples.generated.OrderProto;
+import dev.restate.sdk.examples.types.Status;
+
+public class TypeUtils {
+
+    public static OrderProto.OrderStatus statusToProto(String id, Status status){
+        return OrderProto.OrderStatus.newBuilder().setOrderId(id).setStatus(OrderProto.Status.valueOf(status.name())).build();
+    }
+
+
+}
