+++ java/food-ordering/app/restaurant/src/main/resources/log4j2.properties
@@ -1,0 +1,12 @@
+# Set to debug or trace if log4j initialization is failing
+status = warn
+
+# Console appender configuration
+appender.console.type = Console
+appender.console.name = consoleLogger
+appender.console.layout.type = PatternLayout
+appender.console.layout.pattern = %d{yyyy-MM-dd HH:mm:ss} %-5p %notEmpty{[%X{restateServiceMethod}]}%notEmpty{[%X{restateInvocationId}]} %c - %m%n
+
+# Root logger
+rootLogger.level = info
+rootLogger.appenderRef.stdout.ref = consoleLogger
