+++ java/food-ordering/app/settings.gradle.kts
@@ -1,0 +1,6 @@
+rootProject.name = "food-ordering-app"
+include("restaurant", "restate-app")
+
+dependencyResolutionManagement {
+    repositories { mavenCentral() }
+}
