--- java/food-ordering/app/src/main/java/dev/restate/sdk/examples/types/DeliveryStatus.java
@@ -1,13 +1,0 @@
-package dev.restate.sdk.examples.types;
-
-import dev.restate.sdk.examples.generated.OrderProto;
-
-public class DeliveryStatus {
-    public OrderProto.AssignDeliveryRequest delivery;
-    public boolean pickedUp;
-
-    public DeliveryStatus(OrderProto.AssignDeliveryRequest delivery, boolean pickedUp) {
-        this.delivery = delivery;
-        this.pickedUp = pickedUp;
-    }
-}
