+++ java/food-ordering/scripts/register
@@ -2,6 +2,8 @@
 
 curl -X POST 'localhost:9070/endpoints' -H 'content-type: application/json' -d '{"uri": "http://localhost:9080"}'
 curl -X POST 'localhost:9070/subscriptions' -H 'content-type: application/json' -d '{ "source":"kafka://my-cluster/orders", "sink":"service://order.OrderService/HandleEvent" }'
+curl -X POST 'localhost:9070/subscriptions' -H 'content-type: application/json' -d '{ "source":"kafka://my-cluster/driver-updates", "sink":"service://order.DriverService/UpdateCoordinate" }'
+
 
 curl -X POST -H 'content-type: application/json' localhost:8080/order.DriverSimService/StartDriver -d '{"id": "driver-A"}'
-curl -X POST -H 'content-type: application/json' localhost:8080/order.DriverSimService/StartDriver -d '{"id": "driver-B"}'
+#curl -X POST -H 'content-type: application/json' localhost:8080/order.DriverSimService/StartDriver -d '{"id": "driver-B"}'
